Add YouTube, Spotify, and Twitter data retrieval functions; refactor Analysis.ts to utilize new metadata services for enhanced content analysis
