+++ app/services/ai/Analysis.ts
@@ -1,6 +1,8 @@
 import { ContentType } from "@/app/store/contentState";
 import { GoogleGenAI } from "@google/genai";
-import axios from "axios";
+import { getYoutubeVideoData } from "../metadata/YoutubeVideoData";
+import { getSpotifyData } from "../metadata/SpotifyData";
+import { getTwitterData } from "../metadata/TwitterData";
 
 const ai = new GoogleGenAI({ apiKey: process.env.GEMINI_API_KEY });
 
@@ -57,116 +59,4 @@ export async function analyzeContent(
     .trim();
 
   return JSON.parse(cleanedText);
-}
-
-async function getYoutubeVideoData(url: string) {
-  const videoUrl = new URL(url);
-  let videoId: string | null = null;
-
-  if (videoUrl.hostname === "youtu.be") {
-    videoId = videoUrl.pathname.slice(1);
-  } else if (videoUrl.hostname.includes("youtube.com")) {
-    videoId = videoUrl.searchParams.get("v");
-  }
-
-  if (!videoId) {
-    throw new Error("Could not extract video ID from URL");
-  }
-
-  const response = await axios.get(
-    `https://www.googleapis.com/youtube/v3/videos?id=${videoId}&key=${process.env.YOUTUBE_API_KEY}&part=snippet`
-  );
-
-  const videoData = response.data.items[0].snippet;
-  return {
-    title: videoData.title,
-    description: videoData.description,
-    tags: videoData.tags || [],
-  };
-}
-
-async function getSpotifyUserToken() {
-  const clientId = process.env.SPOTIFY_CLIENT_ID;
-  const clientSecret = process.env.SPOTIFY_CLIENT_SECRET;
-  const auth = Buffer.from(`${clientId}:${clientSecret}`).toString("base64");
-
-  const response = await axios.post(
-    "https://accounts.spotify.com/api/token",
-    new URLSearchParams({ grant_type: "client_credentials" }).toString(),
-    {
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-        Authorization: `Basic ${auth}`,
-      },
-    }
-  );
-
-  return response.data.access_token;
-}
-
-async function getSpotifyData(url: string) {
-  const id = url.split("/").pop()?.split("?")[0];
-  const type = url.includes("/track/") ? "track" : "playlist";
-
-  if (!id) throw new Error("Invalid Spotify URL");
-
-  const token = await getSpotifyUserToken();
-  try {
-    const response = await axios.get(
-      `https://api.spotify.com/v1/${type}s/${id}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    return {
-      title: response.data.name,
-      description:
-        type === "track"
-          ? `By ${response.data.artists[0].name}`
-          : `Playlist by ${response.data.owner.display_name}`,
-      tags: [
-        response.data.type,
-        response.data.artists?.[0]?.name || "Various Artists",
-      ],
-    };
-  } catch (error) {
-    console.error("Failed while fetching metadata", error);
-    throw error;
-  }
-}
-
-async function getTwitterData(url: string) {
-  const tweetId = url.split("/status/")[1]?.split("?")[0];
-  if (!tweetId) {
-    throw new Error("Invalid Twitter URL");
-  }
-  const apiKey = process.env.TWITTER_API_KEY;
-  try {
-    const response = await axios.get(
-      `https://api.twitter.com/2/tweets/${tweetId}`,
-      {
-        headers: {
-          Authorization: `Bearer ${apiKey}`,
-        },
-        params: {
-          "user.fields": "username",
-          expansions: "author_id",
-        },
-      }
-    );
-    const tweet = response.data.data;
-    const user = response.data.includes.users[0];
-
-    return {
-      title: `Tweet by @${user.username}`,
-      description: tweet.text,
-      tags: ["Twitter", user.username, ...(tweet.text.match(/#\w+/g) || [])],
-    };
-  } catch (error) {
-    console.error("Failed to fetch Twitter data", error);
-    throw error;
-  }
 }
